Test / survey and record builders: code cleanup
