--- src/test/recordBuilder/index.ts
@@ -1,1 +1,0 @@
-export { record, entity, attribute } from './recordBuilder'
