--- src/test/surveyBuilder/index.ts
@@ -1,1 +1,0 @@
-export { survey, entity, attribute } from './surveyBuilder'
